refact: TagInput 부분 다시 EventTag.tsx에 합치기
